Further cleanup, things are starting to shape up, considerably!
No more warnings, so we can remove the filtering in the Makefile.

After querying both John and Freek over email about licensing, I
can now release it under the ISC License.  See the excellent
Wikipedia article for details.

   http://en.wikipedia.org/wiki/ISC_license

Apparently John is the original author and Freek is the initiator
and this is now official in the AUTHORS file.

The new README file is currently John's home page for his 1989
Tetris IOCCC obfuscation entry.