Fix level, so that it actually increases.
According to Wikipedia[1]; "With every ten lines that are cleared, the game
enters a new level. As the game progresses, each level causes the
tetrominoes to fall faster...".  The latter part is still TODO.

The score is now also separate from the points, similar to how BSD Tetris
does it.

[1] - http://en.wikipedia.org/wiki/Tetris